+++ build.gradle.kts
@@ -1,0 +1,4 @@
+
+plugins {
+    id("com.gtnewhorizons.gtnhconvention")
+}
