+++ jitpack.yml
@@ -1,0 +1,2 @@
+before_install:
+  - ./gradlew setupCIWorkspace
