--- src/main/java/com/aminoglycoside/modernmarkings/base/items/ModRecipes.java
@@ -1,20 +1,0 @@
-package com.aminoglycoside.modernmarkings.base.items;
-
-import com.aminoglycoside.modernmarkings.base.init.ModBlocks;
-import com.aminoglycoside.modernmarkings.base.init.ModItems;
-import com.aminoglycoside.modernmarkings.base.util.Resource;
-import net.minecraft.init.Items;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.fml.common.registry.ForgeRegistries;
-import net.minecraftforge.fml.common.registry.GameRegistry;
-import net.minecraftforge.oredict.OreDictionary;
-import net.minecraftforge.oredict.ShapedOreRecipe;
-
-public class ModRecipes {
-        public static void init() {
-
-
-        }
-
-    }
