--- src/main/java/com/aminoglycoside/modernmarkings/base/proxy/ClientProxy.java
@@ -1,12 +1,0 @@
-package com.aminoglycoside.modernmarkings.base.proxy;
-
-import net.minecraft.client.renderer.block.model.ModelResourceLocation;
-import net.minecraft.item.Item;
-import net.minecraftforge.client.model.ModelLoader;
-
-public class ClientProxy extends CommonProxy {
-
-    public void registerItemRenderer(Item item, int meta, String id) {
-        ModelLoader.setCustomModelResourceLocation(item, meta, new ModelResourceLocation(item.getRegistryName(), id));
-    }
-}
