--- src/main/java/com/aminoglycoside/modernmarkings/base/proxy/CommonProxy.java
@@ -1,9 +1,0 @@
-package com.aminoglycoside.modernmarkings.base.proxy;
-
-import net.minecraft.item.Item;
-
-public class CommonProxy {
-
-    public void registerItemRenderer(Item item, int meta, String id) {
-    }
-}
