--- src/main/java/com/aminoglycoside/modernmarkings/base/util/IHasModel.java
@@ -1,6 +1,0 @@
-package com.aminoglycoside.modernmarkings.base.util;
-
-public interface IHasModel {
-    // little utility to help us registering all models
-    public void registerModels();
-}
