--- src/main/java/com/aminoglycoside/modernmarkings/base/util/handlers/RegistryHandler.java
@@ -1,48 +1,0 @@
-package com.aminoglycoside.modernmarkings.base.util.handlers;
-
-import com.aminoglycoside.modernmarkings.base.init.ModBlocks;
-import com.aminoglycoside.modernmarkings.base.init.ModItems;
-import com.aminoglycoside.modernmarkings.base.util.IHasModel;
-import net.minecraft.block.Block;
-import net.minecraft.item.Item;
-import net.minecraftforge.client.event.ModelRegistryEvent;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-
-@EventBusSubscriber
-public class RegistryHandler {
-
-    /**
-     * Function to register all Mod Items
-     */
-    @SubscribeEvent
-    public static void onItemRegister(RegistryEvent.Register<Item> event) {
-        event.getRegistry().registerAll(ModItems.ITEMS.toArray(new Item[0]));
-    }
-
-    /**
-     * Function to register all Mod Blocks
-     */
-    @SubscribeEvent
-    public static void onBlockRegister(RegistryEvent.Register<Block> event) {
-        event.getRegistry().registerAll(ModBlocks.BLOCKS.toArray(new Block[0]));
-    }
-
-    /**
-     * Registering the Models for Mod Items and Mod Blocks
-     */
-    @SubscribeEvent
-    public static void onModelRegister(ModelRegistryEvent event) {
-        for (Item item : ModItems.ITEMS) {
-            if (item instanceof IHasModel) {
-                ((IHasModel) item).registerModels();
-            }
-        }
-        for (Block block : ModBlocks.BLOCKS) {
-            if (block instanceof IHasModel) {
-                ((IHasModel) block).registerModels();
-            }
-        }
-    }
-}
