+++ src/main/java/com/myname/mymodid/ClientProxy.java
@@ -1,0 +1,8 @@
+package com.myname.mymodid;
+
+public class ClientProxy extends CommonProxy {
+
+    // Override CommonProxy methods here, if you want a different behaviour on the client (e.g. registering renders).
+    // Don't forget to call the super methods as well.
+
+}
