+++ src/main/java/com/myname/mymodid/CommonProxy.java
@@ -1,0 +1,27 @@
+package com.myname.mymodid;
+
+import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.event.FMLPostInitializationEvent;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.event.FMLServerStartingEvent;
+
+public class CommonProxy {
+
+    // preInit "Run before anything else. Read your config, create blocks, items, etc, and register them with the
+    // GameRegistry." (Remove if not needed)
+    public void preInit(FMLPreInitializationEvent event) {
+        Config.synchronizeConfiguration(event.getSuggestedConfigurationFile());
+
+        MyMod.LOG.info(Config.greeting);
+        MyMod.LOG.info("I am MyMod at version " + Tags.VERSION);
+    }
+
+    // load "Do your mod setup. Build whatever data structures you care about. Register recipes." (Remove if not needed)
+    public void init(FMLInitializationEvent event) {}
+
+    // postInit "Handle interaction with other mods, complete your setup based on this." (Remove if not needed)
+    public void postInit(FMLPostInitializationEvent event) {}
+
+    // register server commands in this event handler (Remove if not needed)
+    public void serverStarting(FMLServerStartingEvent event) {}
+}
