+++ src/main/java/com/myname/mymodid/Config.java
@@ -1,0 +1,20 @@
+package com.myname.mymodid;
+
+import java.io.File;
+
+import net.minecraftforge.common.config.Configuration;
+
+public class Config {
+
+    public static String greeting = "Hello World";
+
+    public static void synchronizeConfiguration(File configFile) {
+        Configuration configuration = new Configuration(configFile);
+
+        greeting = configuration.getString("greeting", Configuration.CATEGORY_GENERAL, greeting, "How shall I greet?");
+
+        if (configuration.hasChanged()) {
+            configuration.save();
+        }
+    }
+}
