+++ src/main/java/modernmarkings/base/items/ModRecipes.java
@@ -1,0 +1,20 @@
+package modernmarkings.base.items;
+
+import com.aminoglycoside.modernmarkings.base.init.ModBlocks;
+import com.aminoglycoside.modernmarkings.base.init.ModItems;
+import com.aminoglycoside.modernmarkings.base.util.Resource;
+import net.minecraft.init.Items;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.fml.common.registry.ForgeRegistries;
+import net.minecraftforge.fml.common.registry.GameRegistry;
+import net.minecraftforge.oredict.OreDictionary;
+import net.minecraftforge.oredict.ShapedOreRecipe;
+
+public class ModRecipes {
+        public static void init() {
+
+
+        }
+
+    }
