+++ src/main/java/modernmarkings/base/proxy/CommonProxy.java
@@ -1,0 +1,9 @@
+package modernmarkings.base.proxy;
+
+import net.minecraft.item.Item;
+
+public class CommonProxy {
+
+    public void registerItemRenderer(Item item, int meta, String id) {
+    }
+}
