+++ src/main/java/modernmarkings/base/util/IHasModel.java
@@ -1,0 +1,6 @@
+package modernmarkings.base.util;
+
+public interface IHasModel {
+    // little utility to help us registering all models
+    public void registerModels();
+}
