+++ src/main/java/modernmarkings/base/util/Resource.java
@@ -1,0 +1,12 @@
+package modernmarkings.base.util;
+
+public class Resource {
+
+    // Change the modid here (it is needed for every .json in the resource folder and for much more!
+    public static final String MOD_ID = "ags_modernmarkings";
+    public static final String NAME = "AGS ModernMarkings Mod";
+    public static final String VERSION = "0.2";
+    public static final String ACEPTED_VERSIONS = "[1.12.2]";
+    public static final String CLIENT_PROXY_CLASS = "com.aminoglycoside.modernmarkings.base.proxy.ClientProxy";
+    public static final String COMMON_PROXY_CLASS = "com..aminoglycoside.modernmarkings.base.proxy.CommonProxy";
+}
